:sparkles: v3 (feature): initial support for binding
