+++ binder/query.go
@@ -1,0 +1,49 @@
+package binder
+
+import (
+	"reflect"
+	"strings"
+
+	"github.com/gofiber/fiber/v3/utils"
+	"github.com/valyala/fasthttp"
+)
+
+type queryBinding struct{}
+
+func (queryBinding) Name() string {
+	return "query"
+}
+
+func (b queryBinding) Bind(reqCtx *fasthttp.RequestCtx, out any) error {
+	data := make(map[string][]string)
+	var err error
+
+	reqCtx.QueryArgs().VisitAll(func(key, val []byte) {
+		if err != nil {
+			return
+		}
+
+		k := utils.UnsafeString(key)
+		v := utils.UnsafeString(val)
+
+		if strings.Contains(k, "[") {
+			k, err = parseParamSquareBrackets(k)
+		}
+
+		if strings.Contains(v, ",") && equalFieldType(out, reflect.Slice, k) {
+			values := strings.Split(v, ",")
+			for i := 0; i < len(values); i++ {
+				data[k] = append(data[k], values[i])
+			}
+		} else {
+			data[k] = append(data[k], v)
+		}
+
+	})
+
+	if err != nil {
+		return err
+	}
+
+	return parseToStruct(b.Name(), out, data)
+}
